+++ src/Classes/GridOpenedManager.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * GRID for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2024 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-grid
@@ -17,18 +17,16 @@ namespace WEM\GridBundle\Classes;
 use Contao\ContentModel;
 use Contao\Database\Result as DbResult;
 use Contao\System;
-use Exception;
-use InvalidArgumentException;
 use WEM\GridBundle\Elements\GridStart as GridStartElement;
 use WEM\GridBundle\Helper\GridBuilder;
 
 class GridOpenedManager
 {
-    /** @var int */
-    protected $level = 0;
+    protected int $level = 0;
+
     private static $instance;
-    /** @var GridBuilder */
-    private $gridBuilder;
+
+    private GridBuilder $gridBuilder;
 
     private function __construct()
     {
@@ -73,9 +71,11 @@ class GridOpenedManager
             ->setLevel($this->level)
             ->setMode($element->grid_mode)
         ;
-        if ('BE' === TL_MODE) {
+        $scopeMatcher = System::getContainer()->get('wem.scope_matcher');
+        if ($scopeMatcher->isBackend()) {
             $grid->addItemClassesForAllResolution('be_item_grid helper');
         }
+
         if (!empty($element->cssID[1])) {
             $grid->addWrapperClasses($element->cssID[1]);
         }
@@ -130,9 +130,8 @@ class GridOpenedManager
         ) {
             return null;
         }
-        end($GLOBALS['WEM']['GRID']);
 
-        $key = key($GLOBALS['WEM']['GRID']);
+        $key = array_key_last($GLOBALS['WEM']['GRID']);
 
         reset($GLOBALS['WEM']['GRID']);
 
@@ -152,6 +151,7 @@ class GridOpenedManager
         ) {
             return null;
         }
+
         $gridsCopy = $GLOBALS['WEM']['GRID'];
         array_pop($gridsCopy);
 
@@ -175,12 +175,11 @@ class GridOpenedManager
         ) {
             return null;
         }
+
         $gridsCopy = $GLOBALS['WEM']['GRID'];
         array_pop($gridsCopy);
 
-        end($gridsCopy);
-
-        $key = key($gridsCopy);
+        $key = array_key_last($gridsCopy);
 
         unset($gridsCopy);
 
@@ -192,7 +191,7 @@ class GridOpenedManager
      *
      * @param string $id The id
      *
-     * @throws Exception if no grid is found
+     * @throws \Exception if no grid is found
      *
      * @return GridOpened the grid
      */
@@ -202,7 +201,7 @@ class GridOpenedManager
             || !\array_key_exists('GRID', $GLOBALS['WEM'])
             || !\array_key_exists($id, $GLOBALS['WEM']['GRID'])
         ) {
-            throw new Exception('The grid doesn\'t exists.');
+            throw new \Exception("The grid doesn't exists.");
         }
 
         return $GLOBALS['WEM']['GRID'][$id];
@@ -218,7 +217,7 @@ class GridOpenedManager
     public function getParentGrid(ContentModel $element): ?GridOpened
     {
         $arrGrid = null;
-        foreach ($GLOBALS['WEM']['GRID'] as $k => $g) {
+        foreach ($GLOBALS['WEM']['GRID'] as $g) {
             if ($g->hasChildByItemId((string) $element->id)) {
                 $arrGrid = $g;
                 break;
@@ -233,14 +232,20 @@ class GridOpenedManager
      *
      * @param ContentModel|DbResult $element The element to check
      *
-     * @throws InvalidArgumentException if the element is not a grid start
+     * @throws \InvalidArgumentException if the element is not a grid start
      */
     public function validateElementAsAGridStart($element): void
     {
-        if (!(is_a($element, DbResult::class) || is_a($element, ContentModel::class) || is_a($element, GridStartElement::class))
+        if (!(
+            is_a($element, DbResult::class)
+            || is_a($element, ContentModel::class)
+            || is_a($element, GridStartElement::class)
+            // || $element instanceof ContentModel::class
+            // || $element instanceof GridStartElement::class
+        )
             || 'grid-start' !== $element->type
         ) {
-            throw new InvalidArgumentException('The element "'.\get_class($element).'" is not a "grid-start"');
+            throw new \InvalidArgumentException('The element "'.\get_class($element).'" is not a "grid-start"');
         }
     }