+++ src/Elements/GridStart.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * GRID for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2024 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-grid
@@ -18,8 +18,8 @@ use Contao\BackendTemplate;
 use Contao\ContentElement;
 use Contao\Database;
 use Contao\System;
+use WEM\GridBundle\Classes\GridCssClassesInheritance;
 use WEM\GridBundle\Classes\GridOpenedManager;
-use WEM\GridBundle\Helper\GridBuilder;
 
 /**
  * Content Element "grid-start".
@@ -27,7 +27,9 @@ use WEM\GridBundle\Helper\GridBuilder;
 class GridStart extends ContentElement
 {
     public const MODE_CUSTOM = 'custom';
+
     public const MODE_AUTOMATIC = 'automatic';
+
     /**
      * Template.
      *
@@ -41,7 +43,8 @@ class GridStart extends ContentElement
     protected function compile(): void
     {
         // Backend template
-        if (TL_MODE === 'BE' && !$this->isForGridElementWizard) {
+        $scopeMatcher = System::getContainer()->get('wem.scope_matcher');
+        if ($scopeMatcher->isBackend() && !$this->isForGridElementWizard) {
             $this->strTemplate = 'be_wildcard';
             $this->Template = new BackendTemplate($this->strTemplate);
             $this->Template->title = $GLOBALS['TL_LANG']['CTE'][$this->type][1];
@@ -59,11 +62,12 @@ class GridStart extends ContentElement
                 $breakpoints = [];
                 // $arrGridValues = GridBuilder::getWrapperClasses($this);
                 $arrGridValues = System::getContainer()->get('wem.grid.helper.grid_builder')->getWrapperClasses($this);
-                foreach ($arrGridValues as $k => $b) {
+                foreach ($arrGridValues as $b) {
                     $b = explode('-', $b);
                     if ('cols' !== $b[0]) {
                         continue;
                     }
+
                     if (2 === \count($b)) {
                         $breakpoint = $this->getBreakpointData('all');
                         $val = $b[1];
@@ -76,6 +80,7 @@ class GridStart extends ContentElement
                         $breakpoints[] = $breakpoint['label'].': '.sprintf($GLOBALS['TL_LANG']['WEM']['GRID']['BE']['nbColsOptionLabel'], $val);
                     }
                 }
+
                 $this->Template->wildcard = 'Config: '.implode(', ', $breakpoints);
             } else {
                 $this->Template->wildcard = 'Config: '.$GLOBALS['TL_LANG']['tl_content']['grid_mode']['automatic'];
@@ -93,7 +98,7 @@ class GridStart extends ContentElement
         $gop = GridOpenedManager::getInstance();
         try {
             $arrGrid = $gop->getGridById((string) $this->id);
-        } catch (\Exception $e) {
+        } catch (\Exception $exception) {
             $gop->openGrid($this);
             $arrGrid = $gop->getGridById((string) $this->id);
         }
@@ -104,7 +109,8 @@ class GridStart extends ContentElement
         } else {
             $this->objModel->classes = $arrGrid->getWrapperClasses();
         }
-        $gridCssClassesInheritance = new \WEM\GridBundle\Classes\GridCssClassesInheritance();
+
+        $gridCssClassesInheritance = new GridCssClassesInheritance();
         $this->objModel->classes = explode(' ', $gridCssClassesInheritance->cleanForFrontendDisplay(implode(' ', $arrGrid->getWrapperClasses())));
 
         // Send the grid_id to template
@@ -118,5 +124,6 @@ class GridStart extends ContentElement
                 return $b;
             }
         }
+        return null;
     }
 }