+++ src/Helper/GridBuilder.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * GRID for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2024 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-grid
@@ -16,7 +16,10 @@ namespace WEM\GridBundle\Helper;
 
 use Contao\ContentModel;
 use Contao\Database\Result as DatabaseResult;
+use Contao\StringUtil;
+use Contao\System;
 use WEM\GridBundle\Classes\GridOpenedManager;
+use WEM\GridBundle\Elements\GridStart;
 
 /**
  * Function to centralize generic code to.
@@ -24,30 +27,34 @@ use WEM\GridBundle\Classes\GridOpenedManager;
 class GridBuilder
 {
     /**
-     * Generate wrapper classes, depending of the element.
+     * Returns an array of CSS classes for the wrapper element of a grid.
      *
-     * @param \ContentModel $objElement [description]
+     * @param ContentModel|DatabaseResult $objElement the content element object
      *
-     * @return [type] [description]
+     * @throws \Exception
+     *
+     * @return array the array of CSS classes
      */
-    public static function getWrapperClasses($objElement)
+    public static function getWrapperClasses($objElement): array
     {
         $arrClasses = [];
 
+        $scopeMatcher = System::getContainer()->get('wem.scope_matcher');
+
         if (!\is_array($objElement->grid_rows)) {
-            $rows = \Contao\StringUtil::deserialize($objElement->grid_rows);
+            $rows = StringUtil::deserialize($objElement->grid_rows);
         } else {
             $rows = $objElement->grid_rows;
         }
 
         if (!\is_array($objElement->grid_cols)) {
-            $cols = \Contao\StringUtil::deserialize($objElement->grid_cols);
+            $cols = StringUtil::deserialize($objElement->grid_cols);
         } else {
             $cols = $objElement->grid_cols;
         }
 
         if (!\is_array($objElement->grid_gap)) {
-            $gap = \Contao\StringUtil::deserialize($objElement->grid_gap);
+            $gap = StringUtil::deserialize($objElement->grid_gap);
         } else {
             $gap = $objElement->grid_gap;
         }
@@ -58,21 +65,19 @@ class GridBuilder
         // }
         $arrClasses[] = 'd-grid';
 
-        if (\WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $objElement->grid_mode) {
+        if (GridStart::MODE_AUTOMATIC === $objElement->grid_mode) {
             $arrClasses[] = 'cols-autofit';
-        } elseif (\WEM\GridBundle\Elements\GridStart::MODE_CUSTOM === $objElement->grid_mode) {
+        } elseif (GridStart::MODE_CUSTOM === $objElement->grid_mode) {
             foreach ($cols as $k => $col) {
                 // Quickfix : we need the first col to be generic, no matter what is the breakpoint
                 if (0 === $k) {
                     $arrClasses[] = sprintf('cols-%d', $col['value']);
-                } else {
-                    if ('FE' === TL_MODE) {
-                        if (0 !== (int) $col['value']) {
-                            $arrClasses[] = sprintf('cols-%s-%d', $col['key'], $col['value']);
-                        }
-                    } else {
+                } elseif ($scopeMatcher->isFrontend()) {
+                    if (0 !== (int) $col['value']) {
                         $arrClasses[] = sprintf('cols-%s-%d', $col['key'], $col['value']);
                     }
+                } else {
+                    $arrClasses[] = sprintf('cols-%s-%d', $col['key'], $col['value']);
                 }
             }
 
@@ -81,14 +86,12 @@ class GridBuilder
                     // Quickfix : we need the first col to be generic, no matter what is the breakpoint
                     if (0 === $k) {
                         $arrClasses[] = sprintf('rows-%d', $row['value']);
-                    } else {
-                        if ('FE' === TL_MODE) {
-                            if (0 !== (int) $row['value']) {
-                                $arrClasses[] = sprintf('rows-%s-%d', $row['key'], $row['value']);
-                            }
-                        } else {
+                    } elseif ($scopeMatcher->isFrontend()) {
+                        if (0 !== (int) $row['value']) {
                             $arrClasses[] = sprintf('rows-%s-%d', $row['key'], $row['value']);
                         }
+                    } else {
+                        $arrClasses[] = sprintf('rows-%s-%d', $row['key'], $row['value']);
                     }
                 }
             }
@@ -102,9 +105,12 @@ class GridBuilder
     }
 
     /**
-     * Generate item classes, depending of the element.
+     * Generate item classes, depending on the element.
      *
-     * @param ContentModel|DatabaseResult $objElement [description]
+     * @param ContentModel|DatabaseResult $objElement the grid element object
+     * @param bool|null                   $forForm    if true, prepares the classes for rendering in a form, for rendering in the grid
+     *
+     * @return array an array of classes for the grid item element
      */
     public static function getItemClasses($objElement, ?bool $forForm = false): array
     {
@@ -112,7 +118,7 @@ class GridBuilder
         if (\is_array($objElement->grid_cols)) {
             $cols = $objElement->grid_cols;
         } else {
-            $cols = \Contao\StringUtil::deserialize($objElement->grid_cols);
+            $cols = StringUtil::deserialize($objElement->grid_cols);
         }
 
         $arrClasses[] = 'item-grid';
@@ -121,10 +127,11 @@ class GridBuilder
         $arrItemsClasses = [];
         if (null !== $objElement->grid_items) {
             if (!\is_array($objElement->grid_items)) {
-                $items = \Contao\StringUtil::deserialize($objElement->grid_items);
+                $items = StringUtil::deserialize($objElement->grid_items);
             } else {
                 $items = $objElement->grid_items;
             }
+
             if (!$forForm) {
                 foreach ($items as $itemId => $classes) {
                     if (\is_array($classes)) {
@@ -145,6 +152,8 @@ class GridBuilder
      * Returns a "fake" grid element to allow element to be placed at the beggining of the grid.
      *
      * @param string $gridId The grid's id
+     *
+     * @throws \Exception
      */
     public function fakeFirstGridElementMarkup(string $gridId): string
     {
@@ -153,13 +162,15 @@ class GridBuilder
         $grid = $gop->getGridById($gridId);
 
         // $additionnalCssClasses = \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-',$grid->getWrapperClasses()[1]);
-        $additionnalCssClasses = \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-', implode(' ', $grid->getWrapperColsClassesWithoutResolutionSpecificClasses()));
+        $additionnalCssClasses = GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-', implode(' ', $grid->getWrapperColsClassesWithoutResolutionSpecificClasses()));
 
         return sprintf('<div class="item-grid be_item_grid fake-helper be_item_grid_fake %s" dropable="true" draggable="false" data-type="fake-first-element">%s</div>', $additionnalCssClasses, $GLOBALS['TL_LANG']['WEM']['GRID']['BE']['placeToGridStart']);
     }
 
     /**
      * Returns a "fake" grid element to allow element to be placed at the end of the grid.
+     *
+     * @throws \Exception
      */
     public function fakeLastGridElementMarkup(string $gridId): string
     {
@@ -167,22 +178,24 @@ class GridBuilder
 
         $grid = $gop->getGridById($gridId);
 
-        // $additionnalCssClasses = \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-',$grid->getWrapperClasses()[1]);
-        $additionnalCssClasses = \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-', implode(' ', $grid->getWrapperColsClassesWithoutResolutionSpecificClasses()));
+        // $additionnalCssClasses = GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-',$grid->getWrapperClasses()[1]);
+        $additionnalCssClasses = GridStart::MODE_AUTOMATIC === $grid->getMode() ? 'cols-span-all' : str_replace('cols-', 'cols-span-', implode(' ', $grid->getWrapperColsClassesWithoutResolutionSpecificClasses()));
 
         return sprintf('<div class="item-grid be_item_grid fake-helper be_item_grid_fake %s" dropable="true" draggable="false" data-type="fake-last-element">%s</div>', $additionnalCssClasses, $GLOBALS['TL_LANG']['WEM']['GRID']['BE']['placeToGridEnd']);
     }
 
     /**
      * Returns a "fake" grid element to allow new elements to be added at the end of the grid.
+     *
+     * @throws \Exception
      */
-    public function fakeNewGridElementMarkup(string $gridId): string
+    public function fakeNewGridElementMarkup(string $gridId): string // TODO : One day, delete all the function because she is useless.
     {
         $gop = GridOpenedManager::getInstance();
 
         $grid = $gop->getGridById($gridId);
 
-        $additionnalCssClasses = \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC === $grid->getMode() ? '' : '';
+        $additionnalCssClasses = GridStart::MODE_AUTOMATIC === $grid->getMode() ? '' : '';
 
         return sprintf('<div class="item-grid be_item_grid fake-helper be_item_grid_fake %s" dropable="false" draggable="false" data-type="fake-new-element"><div class="item-new"></div></div>', $additionnalCssClasses);
     }