+++ src/Resources/contao/config/config.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * GRID for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2024 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-grid
@@ -12,15 +12,19 @@ declare(strict_types=1);
  * @link     https://github.com/Web-Ex-Machina/contao-grid/
  */
 
+use Contao\ArrayUtil;
+use WEM\GridBundle\Elements;
+use WEM\GridBundle\Widgets;
+
 // Add the Grid Wrapper Content Element
-array_insert(
+ArrayUtil::arrayInsert(
     $GLOBALS['TL_CTE'],
     \count($GLOBALS['TL_CTE']) + 1,
     [
         'grid' => [
-            'grid-start' => 'WEM\GridBundle\Elements\GridStart',
-            'grid-stop' => 'WEM\GridBundle\Elements\GridStop',
-            'grid-item-empty' => 'WEM\GridBundle\Elements\GridItemEmpty',
+            'grid-start' => Elements\GridStart::class,
+            'grid-stop' => Elements\GridStop::class,
+            'grid-item-empty' => Elements\GridItemEmpty::class,
         ],
     ]
 );
@@ -36,6 +40,6 @@ $GLOBALS['TL_WRAPPERS']['stop'][] = 'grid-stop';
 $GLOBALS['TL_HOOKS']['getContentElement'][] = ['wem.grid.event_listener.get_content_element', '__invoke'];
 
 // Add Backend Wizard
-$GLOBALS['BE_FFL']['gridElementWizard'] = 'WEM\GridBundle\Widgets\GridElementWizard';
-$GLOBALS['BE_FFL']['gridBreakpointsValuesWizard'] = 'WEM\GridBundle\Widgets\GridBreakpointsValuesWizard';
-$GLOBALS['BE_FFL']['gridGapValuesWizard'] = 'WEM\GridBundle\Widgets\GridGapValuesWizard';
+$GLOBALS['BE_FFL']['gridElementWizard'] = Widgets\GridElementWizard::class;
+$GLOBALS['BE_FFL']['gridBreakpointsValuesWizard'] = Widgets\GridBreakpointsValuesWizard::class;
+$GLOBALS['BE_FFL']['gridGapValuesWizard'] = Widgets\GridGapValuesWizard::class;