+++ src/Resources/contao/dca/tl_content.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 /**
  * GRID for Contao Open Source CMS
- * Copyright (c) 2015-2022 Web ex Machina
+ * Copyright (c) 2015-2024 Web ex Machina
  *
  * @category ContaoBundle
  * @package  Web-Ex-Machina/contao-grid
@@ -26,8 +26,8 @@ $GLOBALS['TL_DCA']['tl_content']['palettes']['grid-start'] = '{type_legend},type
 $GLOBALS['TL_DCA']['tl_content']['palettes']['grid-stop'] = '{type_legend},type;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID;{invisible_legend:hide},invisible,start,stop';
 
 // update grid content elements subpalettes
-$GLOBALS['TL_DCA']['tl_content']['subpalettes']['grid_mode_'.\WEM\GridBundle\Elements\GridStart::MODE_CUSTOM] = 'grid_cols';
-$GLOBALS['TL_DCA']['tl_content']['subpalettes']['grid_mode_'.\WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC] = '';
+$GLOBALS['TL_DCA']['tl_content']['subpalettes']['grid_mode_'.WEM\GridBundle\Elements\GridStart::MODE_CUSTOM] = 'grid_cols';
+$GLOBALS['TL_DCA']['tl_content']['subpalettes']['grid_mode_'.WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC] = '';
 
 // Update tl_content fields
 $GLOBALS['TL_DCA']['tl_content']['fields']['grid_row_class'] = [
@@ -66,10 +66,10 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['grid_mode'] = [
     'exclude' => true,
     'inputType' => 'select',
     'options' => [
-        \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC => $GLOBALS['TL_LANG']['tl_content']['grid_mode']['automatic'],
-        \WEM\GridBundle\Elements\GridStart::MODE_CUSTOM => $GLOBALS['TL_LANG']['tl_content']['grid_mode']['custom'],
+        WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC => &$GLOBALS['TL_LANG']['tl_content']['grid_mode']['automatic'],
+        WEM\GridBundle\Elements\GridStart::MODE_CUSTOM => &$GLOBALS['TL_LANG']['tl_content']['grid_mode']['custom'],
     ],
-    'default' => \WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC,
+    'default' => WEM\GridBundle\Elements\GridStart::MODE_AUTOMATIC,
     'eval' => ['submitOnChange' => true, 'tl_class' => 'clr'],
-    'sql' => "varchar(9) NOT NULL default '".\WEM\GridBundle\Elements\GridStart::MODE_CUSTOM."'", // for backward compatibility on all existing grids
+    'sql' => "varchar(9) NOT NULL default '".WEM\GridBundle\Elements\GridStart::MODE_CUSTOM."'", // for backward compatibility on all existing grids
 ];